Make youtube source IMusicSource

+++ Zoom/Song.cs
@@ -1,9 +1,9 @@
 namespace Zoom;
 
 public record Song(SongInfo Info, SoundStream Stream);
-public record SongInfo(string? Title, int LengthSeconds)
+public record SongInfo(string? Title, double LengthSeconds)
 {
-    public static SongInfo Create(string author, string title, int lengthSeconds) => new(GetTitle(author, title), lengthSeconds);
+    public static SongInfo Create(string author, string title, double lengthSeconds) => new(GetTitle(author, title), lengthSeconds);
 
     public override string ToString() => Title ?? "<нет имени>";
